- Moved the api dir to internal/  - Added sql scripts for taxify db
