+++ internal/models/01_drop_db.sql
@@ -1,0 +1,1 @@
+drop database if exists taxify;
