+++ internal/models/02_create_db.sql
@@ -1,0 +1,1 @@
+create database taxify;
