+++ internal/models/03_create_tables.sql
@@ -1,0 +1,6 @@
+create table taxify.users (
+  user_id int unsigned auto_increment,
+  constraint pk_user_id primary key (user_id),
+  phone_number varchar(64) not null unique,
+  verify_sid varchar(255)
+);
