+++ internal/models/04_seed_data.sql
@@ -1,0 +1,3 @@
+insert into taxify.users
+    (user_id, phone_number, verify_sid)
+    VALUES (null, '04023734503', '2906');
